+++ pytorch_pfn_extras/onnx/export_testcase.py
@@ -26,7 +26,7 @@ from pytorch_pfn_extras.onnx.strip_large_tensor import is_large_tensor
 from pytorch_pfn_extras.onnx.strip_large_tensor import _strip_raw_data
 from pytorch_pfn_extras.onnx.strip_large_tensor import \
     _strip_large_initializer_raw_data
-from pytorch_pfn_extras.onnx.export import export as pfto_export
+from pytorch_pfn_extras.onnx.pfto_exporter import export as pfto_export
 
 
 def _model_to_graph_with_value_names(
